moved utils.ts
