+++ src/lib/utils.ts
@@ -1,6 +1,24 @@
-import { clsx, type ClassValue } from "clsx";
-import { twMerge } from "tailwind-merge";
+// Logger utility for error handling and messaging
 
-export function cn(...inputs: ClassValue[]) {
-    return twMerge(clsx(inputs));
+const isDev = import.meta.env.MODE === 'development';
+
+export function logError(error: unknown, context?: string) {
+    if (isDev) {
+        // Detailed error in dev
+        console.error(`[DEV ERROR]${context ? ' [' + context + ']' : ''}`, error);
+    } else {
+        // User-friendly error in prod
+        console.error(`[ERROR]${context ? ' [' + context + ']' : ''}`, error instanceof Error ? error.message : String(error));
+    }
 }
+
+export function getUserErrorMessage(error: unknown): string {
+    if (isDev && error instanceof Error) {
+        return `${error.message}\n${error.stack}`;
+    }
+    if (error instanceof Error) {
+        return error.message;
+    }
+    return 'An unexpected error occurred.';
+}
+
